+++ common/src/main/kotlin/com/rpmtw/rpmtw_platform_mod/translation/machineTranslation/QueueText.kt
@@ -1,5 +1,5 @@
 package com.rpmtw.rpmtw_platform_mod.translation.machineTranslation
 
-import com.rpmtw.rpmtw_platform_mod.translation.language.TranslateLanguage
+import com.rpmtw.rpmtw_platform_mod.translation.TranslateLanguage
 
 data class QueueText(val text: String, val language: TranslateLanguage)
